[testrunner] Increase timeout waiting for a test to start over serial.
Testrunner retries running a test over serial because of a flake
observed where we send a command to run a test and it never starts.
However, because the timeout was too short, it would resend the command
while the first command was already running. This caused tests to be run
multiple times whether they passed or failed. This change does not fix
the attached bug, but it should reduce the chances of it happening.

I also changed MatchingReader.Read() to return an io.EOF when it finds a
match (as opposed to the next time Read() is called). This would make
ReadUntilMatch work more properly. Before, it could enter a case where
it would Read() and find a match but return a nil error, causing it to
continue the loop. However, if the context was cancelled in that time
(i.e due to a timeout), it would break out of the loop, log the last
bytes read and return the context error. This would cause ReadUntilMatch
to return an error even if it found a match and the last bytes read
would show the matching string, which would cause confusion in the logs.

Bug: 68733
Change-Id: I17fad31fc81d740c3825b278d677c3ba729722ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487830
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>