[audio][test] Fix issue with --gtest_repeat
Test suites can set custom HermeticAudioEnvironment variables. Currently
those variables are not reset between tests. We haven't noticed this
problem because our tests happen to be ordered like this:

- Test suites that use the default environment
- Test suites that use custom environments

When these suites are run in order, everything works fine. However, if
these suites are run out-of-order, or if the entire test binary is run
multiple times via --gtest_repeat, the custom HermeticAudioEnvironment
variables will be used for tests that need the default environment.

To fix this problem, this CL adds a default implementation of
SetUpTestSuite which explicitly asks for the default environment.

Change-Id: I01a99b1a83099197bc838ee09ff4932708edd452
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487961
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Payton Bradsky <pbradsky@google.com>