[audio_core] Always flush/invalidate RingBuffers
RingBuffers are synchronized by time, not by conventional
synchronization primitives such as locks. Hence, RingBuffer reads
and writes are not ordered by a conventional happens-before relation.
We must insert fences to ensure memory coherency:

- Writers must flush their cache to ensure that writes are visible
  to other processors and devices (such as the HW speaker).

- Readers must invalidate their cache to ensure they do not have a stale
  copy of the buffer in cache.

The write fence is already done by driver_output.cc (note that
driver_output.cc does not use WritableRingBuffer; instead it makes
raw copies into the ring buffer):
https://cs.opensource.google/fuchsia/fuchsia/+/master:src/media/audio/audio_core/driver_output.cc;drc=8f6bc71409eddd4cedf7e6731bc9e4448218204e;l=285

The read fence is already done for RingBuffers shared with HW.

Hence, the only behavioral change in this CL is to add read and write
fences to SW buffers, i.e. to loopback buffers. There is a chance this
may induce a small performance hit, however it should be small (at most
one per mix period), and in any case, the fences are required for
correctness -- as explained in comments added in this CL, it's likely
that SW buffers need fences rarely because SW buffers are typically
written and read infrequently and in large batches, as opposed to HW
buffers, which are written and read continuously by HW devices.

Fixed: 50442
Change-Id: I67b4ee937f3a92bd168b9c77c59428dd3c3d7cb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487823
Reviewed-by: Payton Bradsky <pbradsky@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>