[stdcompat]: cpp20::span.
To simplify implmentation, we delegate the storage to an internal
struct, so we specialize that intead of the entire span implementation.

This CL just introduces the span construct, there are some constructs
missing such as:

cpp20::byte
cpp20::as_bytes
cpp20::as_writeable_bytes

Test are exhaustive and constexpr to verify that span works as intended
in constexpr context.

There are asserts for each undefined behvior, so bugs can be caught on
debug builds.

Introduced a new binary for testing, LIB_STDCOMPAT_NO_INLINE_VARIABLES
so this code is excersided in more environments, and we can more easily
recreate the conditions of the sdk.

Test: stdcompat-tests
Change-Id: Ic656a07bc05fc84514d30b04e973234eb5785654
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483058
Reviewed-by: Mark Schott <schottm@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>