[zxdb] Fix debug adapter server tests
accept() blocks indefinitely and does not exit if server socket is
closed after it is called. This leads to a race in ~DebugAdaperServer,
when it is trying to shutdown the thread. Using select() and pipe
to signal thread exit to fix this.

Fixed: 66771
Fixed: 66134
Test: fx test zxdb_tests
Change-Id: Id1614f9b0a64e748d601095765fd7c9d32b348ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488209
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>