[events] Extract Events Client Library from OpaqueTest
OpaqueTest is on the path to deprecation. The events client
library is bundled with OpaqueTest but it doesn't need to be.

Extract this library as `component_events` so that OpaqueTest can
be safely deprecated.

`component_manager/testing:test_utils_lib` and `component_events`
are both marked as testonly. This impacts build rules that use
these libraries.

`step` is marked as testonly, otherwise it requires an exception
from these libraries.

Test: fx test hub-integration-test
      fx test storage-integration-test
      ... and more
Change-Id: I44b4950b09985a11aee5c9e2d5d247bcc8fb8b2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475897
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>