[bt][gap] Stop auto-connecting after persistent HCI errors
If we fail to connect to a peer (even after the retry logic) with any
of a set of specific HCI statuses (enumerated in ShouldAlwaysStopAuto-
Connecting), it's likely that there is a persistent issue preventing
the connection. If we have the auto-connect flag set for that peer, it
is possible for us to enter a loop where:
1. We see a connectable advertisement from that peer.
2. Auto-connect is set, so we try to connect to the peer
3. The connection to that peer fails, even though we are still in
   advertising range (e.g. due to asymmetric TX power).
4. GOTO 1.

To prevent this, add a step 3.5 (for auto-connect-initiated connection
attempts only), which sets the auto-connect behavior to not
auto-connect until after the next successful connection.

Bug: b/178810924
Test: `$ fx test bt-host-gap-tests`, added the test
      GAP_LowEnergyConnectionManagerTest.
      ConnectSucceedsThenAutoConnectFailsDisablesAutoConnect, which
      emulates a successful connection+bond, & disconnect, then an
      auto-connect attempt which fails with one of the predefined
      HCI statuses and verifies that bt-host will not auto-connect.
Change-Id: I1165de4bd315751a6ba9258a7608afc930c6cf49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485752
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>
Fuchsia-Auto-Submit: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
(cherry picked from commit 5e1fa50f02522491e9aa7931f69b7a4c06a25b6e)