[codec] Free input packets shortly after stream failure
If a stream has seen a failure, including an async failure
from the CodecAdapter, any further input packets for that
stream are freed back to the client, and any packets already
with CodecAdapter (or racing) are freed either by the
CodecAdapter or by the subsequent
CodecAdapter::CoreCodecStopStream().

Test coverage for this is coming in fxrev.dev/486837.

BUG: fxbug.dev/70521
Change-Id: I689f5bddfe1e63cbda1f4cdefc614eb499129744
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488199
Reviewed-by: RJ Ascani <rjascani@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
API-Review: Dale Sather <dalesat@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>