[codec_impl] Send OnStreamFailed before OnFreeInputPacket
It is largely expected for decrypting StreamProcessors to encounter a
stream failure on the first decrypt packet, as keys may not be available
yet. When this occurs, a OnStreamFailed event is sent and the stream
must be restarted. In http://fxrev.dev/486344, a change was made to
call OnFreeInputPacket for the packet that failed due to the lack of
keys to signal the client that the packet is free. However, since those
are independent events, there is no indication to the client that this
packet failed. In some h264 streams, if that first packet is not
retried, then the subsequent decoding attempt will fail as well.

For that reason, it is temporarily preferable to have the client assume
that any input packets and buffers currently with the StreamProcessor at
the point of stream failure were unable to be completed. The
OnFreeInputPacket events for those pending packets will be sent after
the OnStreamFailed event so that the client knows that they are
available for re-use.

Bug: b/177173519, b/180649772
Change-Id: I7ae80bb6c11b2cd122d22b38606c8c5a342d64b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486837
Reviewed-by: Dustin Green <dustingreen@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Commit-Queue: RJ Ascani <rjascani@google.com>
API-Review: Dale Sather <dalesat@google.com>