[debugger] Hook up thread registers.
The thread registers were not being passed through the unwinder to the
register listings. This meant that TLS values were not available. The
solution is to just copy the thread registers from the OS structure to
every stack frame under the assumption that they never change across
frames.

Adds a missing register record for gsbase on x64. This was causing an
assertion (in addition to meaning it wasn't available).

Bug: 6570, 53940

Change-Id: I10169ccd455aefe70d7f11965e4073edf62568d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489028
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>