[diag_plat] If a component is fully filtered, we can get a null payload w/ no errors.
We should never be able to get a null payload and no errors. Given that a fully filtered
hierarchy should produce a null payload, and such behavior is depended on by tests/clients,
we should instead add error messages to the schema making it clear that the hierrarchy was fully filtered.

Bug: 66756
Change-Id: I686b58c23842684ebb67fa1842d582da24f2ad00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488215
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Luke Nicholson <lukenicholson@google.com>