[brcmfmac] Fix workqueue test flakiness
A workqueue test would sometimes access stack data that was no longer
valid because the test had finished but the access was running on the
default workqueue which had not yet finished. A previous fix was
attempted for this but it did not take into consideration that the
workqueue itself also accesses the WorkItem object after the work
callback has finished.

This fix uses the signaler mechanism of the WorkItem to ensure that the
test method doesn't exit until the work has actually completed.

Bug: 70435
Test: Ran the test
Change-Id: I9c73cd270fe08af19c2c822143b04c2695f203d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487417
Reviewed-by: John Sheu <sheu@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>