[fx set] Deprecate --help-args flag
It would be awkward to support in the new Go version of fx set that's
integrated with fint, because fint is only designed to run `gn gen`, not
arbitrary GN commands.

Also, there's now the alternative of running `fx set` with the desired
args and then `fx args` to view argument documentation.

fx metrics suggest that this flag hasn't been used by anyone in the last
year, so I doubt it will be disruptive to deprecate it and eventually
removing it by leaving it out of the Go version `fx set`.

Also remove mentions of --help-args from the `fx set -h` string, along
with `--build-dir`, which has also been deprecated and will not be
supported by the new `fx-set` command.

Bug: 68465
Change-Id: If247055c4a75b303b35fc816162d6feba47f3daa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489118
Reviewed-by: Gary Boone <gboone@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>