[fidl] Enable handle rights in fidl-format
The motivation for this is that handle rights are appearing in a
number of test files now and if you run "fx format-code" in a
CL that modifies one of these files, it will fail because the
handle files are considered invalid. Further, often prevent
formatting of other files.

It is relatively low risk to enable this flag in the formatter
as it generally won't hide errors unless users were trying to use
handle rights anyways.

Change-Id: Ib59b4faca6de917f08972d0d2ae60b53288b8e15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485439
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>