[time][fake-clock] Use fake time for a timekeeper integ test
Adds a simple integration test that verifies that timekeeper restarts
a time source if it has not produced any samples in an hour.
This test case does not need UTC support from fake-clock, but future
integration tests for timekeeper will require it.

As this is one of the first Rust components using fake-clock, this is a
test of both timekeeper and fake-clock.

Bug: 61500

Multiply: timekeeper-integration

Change-Id: I5fe88ce76bc75a68508b4f09e744f33185a054af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484549
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>