[gigaboot] add support for new partition names
Adds support for `fastboot flash` to specify either new or old partition
names. Getting this in early allows it to soak for a bit so that devs
will flash it to their devices, at which point we could start
transitioning to the new patition scheme if we want.

Bug: 68692
Test: unittests, `fx flash-remote`, `fastboot flash ziron_a <img>`
Change-Id: I928d814c5043493949d98862fe5542f2d9eec728
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490004
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>