[settings][policy] Return early during calculating external volume if dividing by zero
Adds an early return to prevent passing NaN into the round function. At
the moment the code works due to max picking 0 over NaN, but best to
handle this explicitly so that this won't break if the round function
changes. Also adds a test case for this edge case.

Fixed: 70760
Test: fx test setui_service_tests
Change-Id: I24827e611d8db4a7d0bc1db139502ad4a8ee042f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490009
Commit-Queue: William Xiao <wxyz@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>