[fidl][docs] Move source compatibility test READMEs to //docs
All READMEs are moved into a "compatibility guide" section
which is a sibling to the "developer guides". This removes
the top-level separation between "guide"
type documents and "tutorial" type documents, but it
seems worth highlighting this section instead of nesting
it inside the developer guides.

This CL also makes accompanying changes to the gen tool:
* docs are now generated directly to //docs
* There's a separate regen_toc command to generate the toc for
  the "Compatibility guide" section. We don't do this automatically
  to avoid either issues with modifying files or regenerating
  the toc every time a single test changes. Since users will
  need to manually check the ABI/API compat guide for any
  manual changes they need to make as a separate step when
  adding/removing tests, running a separate regen_toc command
  shouldn't be too inconvenient.

Change-Id: I64caf9a86418215268be13b95e314d74b5a0c5b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487458
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>