[kernel][heap] Enable sized delete, assert static size fits in heap block
Enable generation of calls to sized-delete operators; when a compiler
knows that an object being deallocated has a particular size or alignment,
it can call a version of the delete operator with the size specified.

Some allocators can optimize for this path and avoid header lookups;
our kernel heap allocator cannot - its structure would make this
somewhat difficult. Instead, assert that the passed-in size is <= the
size in the malloc block's header. This would guard against certain
corruption of the allocator free list - if we free an object bigger
than its header, we may have underran our buffer.

This assert has very minimal perf costs on x86 - most perfcompare tests
are unaffected, the ones that are are affected marginally.

Change-Id: I4b7efae14d76089dde51cae0aef1aef7c5373461
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466819
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Kostya Kortchinsky <kostyak@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>