[blobfs] ifdef out some unnecessary paging code
When using the new pager, we don't need some of the mechanics of the
Blob class. This puts some of those items behind the new pager ifdef.

The new pager still does not work and is defaulted to off. There should
be no behavior change.

Bug: 51111

Change-Id: If7cf325ab8fc58175d537ecc5d639c002736975b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490862
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>