[async] Add method to delay a promise until and object is signaled
Like the time-based delay methods that were submitted before this, this
method on async::Executor produces a promise that will not complete
until a specified condition, in this case a zx_handle_t being signaled
with the signal provided when the user creates the promise.

Like the previous methods, the waiting starts when the method is called,
so if the user signals the handle before the promise is scheduled, it
will complete immediately when it is scheduled.

Bug: 70802
Test: async-test
Change-Id: Icb4afd8d40897f498d6774245d59869c542ac3a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491438
Commit-Queue: Mark Schott <schottm@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
API-Review: Dale Sather <dalesat@google.com>