[fidl][sanitizers] Remove dependency of zircon/system/ulib on old C bindings
Remove the dependency of debugdata.cc on the deprecated C bindings for
fuchsia.io for calling fuchsia.io.Directory/Open. Unfortunately, because
llcpp depends (in some paths) on new/delete, and this code is used in
libc, this code cannot depend on llcpp as a replacement.

As a not great alternative, and because this is the only place that uses
this protocol that has this restriction, simply copy a few constants and
the structure definition from the C bindings inline to allow the fidl
protocol to be interacted with. A test to confirm that the local copy
does not drift from llcpp is added to z/s/ulib/c/test/debugdata.cc.

Bug: 68648
Change-Id: I5ba3394d103941600cd55e9166afcff66f25a426
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488000
Commit-Queue: Scott Graham <scottmg@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>