[debugger] Enable timeout for symbol downloading
It's suspected that the hang of the symbolizer on infra bots is caused
by the network. Adding a timeout for symbol downloading makes it more
robust. The timeout is set to 10s for connecting and 10s for reading,
which should be sufficient for today's network.

Bug: 66920
Change-Id: I55567a3703db2cc921a17558358907af033fb2f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491121
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>