[media/vnext] new fidl libraries for media vnext prototyping
This CL adds some fidl libraries to //src/media/vnext/fidl. These
definitions are a 'work in progress', but will be checked in so that
members of the media team can write code against them to refine and
expand on them. The intention is that they will eventually move into
//sdk/fidl and become part of the sdk.

fuchsia.media2 has the '2' suffix so that the 'one definition rule'
is not violated. When this becomes part of the sdk, we will need a
transition plan that allows us to replace the current fuchsia.media
definitions.

Definitions in fuchsia.media2 and fuchsia.mediastreams are based on
design documents currently under review. Definitions in fuchsia.audio
have not yet been subject to review and are speculative.

Test: no tests - fidl definitions only

Change-Id: I1f07b55d8ea1e56fd21f4501a27dda0f3487c1f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491578
API-Review: Dale Sather <dalesat@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Reviewed-by: RJ Ascani <rjascani@google.com>