[e2] Deflake ermine shell test take 2
The test to see if Recents is visible needs to take into account
that the terminal view is shifted to the right if Recents is visible.

The test was failing because Recents is a toggle, it can be cancelled
if focus is moved anywhere in the system (and other reasons). Checking
if the view rect is shifted to right, if Recents is visible, is more
robust.

Change-Id: I37abd4c21be94ee5bda7b6b933d365f579a3490e
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/492002
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>