[flatland] add and use PresentArgs, squashable flag
This is the first in a series of CLs to convert Flatland to use Present3
semantics, as defined in http://fxbug.dev/69551.

This change has Flatland use a table, PresentArgs, for Present(), as
opposed to individual fields. This increases the flexibility of
Present(). No field is necessary, and if a field is omitted Flatland
will use a reasonable default value.

This change adds a field to PresentArgs called squashable. Presents with
|squashable| set to false can not be combined with subsequent Presents,
regardless of render timing and presentation times.

Bug: 69551
Fixed: 45945
Test: Unittests added. All unittests pass.

Change-Id: I5aec8deb47725631b5706ba7a50fb06009b6e002
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490721
API-Review: John Bauman <jbauman@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>
Reviewed-by: Mikael Pessa <mikaelpessa@google.com>
Commit-Queue: Felipe Archondo <felipearchondo@google.com>