[pkg-resolver] Add TCP keepalives to HTTP clients
NAT is often coupled with firewalling, and this occasionally performs
connection tracking. Flushing connection tracking storage results in
these firewalls "forgetting" about established connections, requiring
machines inside the firewalled zone to resume these connections. This is
problematic for HTTP/1.1 clients, which do not expect to send data after
the request. Clients blackholed in this fashion must make use of TCP
keep alives to recover from this situation.

This situation was observed to occur in scenarios when toggling the "Use
IPv6" setting on a Nest Wifi AP while pkg-resolver was pulling packages.
Tests were performed by toggling the "enable IPv6" setting while
performing a blob fetch over IPv4 and IPv6. (When using IPv6, the switch
was toggled off and back on.) Performing this action appeared to have
the effect described in the previous paragraph.

Blob fetches over IPv4 timed out in this scenario. With this change,
blob fetches are observed to succeed.

We may be able to automatically test this with network emulation by
inserting black-hole routes. This is a significantly larger effort, and
manual testing is reliable in this case.

TCP keepalive timeouts could trigger concurrently with HTTP timeouts,
and our retry behavior would've differed depending on whether the error
came from hyper or a fetch timeout. This has been remedied such that all
of these cases are eligible for retry, and existing tests have been
fixed.

The auto endpoint updater has been updated to use this new interface
which slightly changes its timeout semantics, but not to an appreciable
degree.

Fixed: 70798
Change-Id: I3b5bffc21ff3687a15748889d5afdf92e2fa11d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491138
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>