[component_manager] Wait for shutdown to mark component deleted
This is a lifecycle behavior change that causes a component to be shut
down before it is marked deleted. ("Marked deleted" means the component
is hidden from the tree, but it may still have resources to be cleaned
up.)

The most visibile side effect is that Realm.DestroyChild will wait for
the child to stop before returning.  Note that with the new behavior, a
child cannot be recreated with the same name until shutdown completes.

This should also yield more intuitive system behavior. Users do not need
to be concerned about "zombie" components that continue to run and have
side effects but are not visible through normal means.

Change-Id: I08af4e2db29db7a1373de51c37ab88c6cff51d06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489417
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Gary Bressler <geb@google.com>