[brcmfmac] Fix re-connection issue after rebooting from WPA3 connection
This change fixed this issue by issuing an BRCMF_C_DISASSOC command
to firmware when an association is rejected or there is an
connection timeout. This is to clean up the connection info in AP of
our device.

Also this change clean up disconnecting status bit when the
BRCMF_C_DISASSOC is failed to send to the firmware.

Adding a delay in policy for WPA3 connection retries is also needed.
This is to bypass the WPA3 protection for re-association spam attack,
and is implemented in fxr/490941.

Bug: 69067
Test:
  - Manually verified with donut, steps:
    1) Save and connect to WPA3 AP.
    2) Reboot the device with "dm reboot", verifies the connection
    status.
  - ACTS test case which covers the same steps.
  - Changed the unittest in sim/test/assoc_test.cc to verify the driver
  behaviour change.

Change-Id: Id13487cb610d898ef5615ca079fc20deed9c3d28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482460
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>