[fidlc] Fix handle rights in coding tables
In the existing version, handle rights are not included in the
disambiguation calculation for handles in coding tables. This led to the
wrong handle type sometimes being pointed to in the coding table.

Additionally, disabled handle rights GIDL tests are enabled in this CL
for LLCPP.

Fixed: 69798
Change-Id: Ibc7b10124a8fe59495735ee6b0ccd3f6457386cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490400
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>