[fidlgen] Fix generation of bits constants
Generated code for bits constants previously wasn't building. The
const.test.fidl and constants.test.fidl files both were disabled for
many of the bindings so these errors were not caught. I created a new
test file that is more basic and focuses on bits constants that can be
enabled across all bindings.

The motivation for this is to be able to reference constants defined in
the zx library for handle rights, e.g. zx.RIGHTS_BASIC.

Change-Id: If69902c454b509970f6e8039bfb81b8b73d0fdb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491319
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>