[io_util] Introduce AsyncReadAt and AsyncGetSize.
AsyncReadAt is similar to the standard AsyncRead trait with an
additional offset, AsyncGetSize is a trait to get file size
asynchronously.

Also included a few util methods in AsyncReadAtExt and AsyncGetSizeExt.

And a wrapper that implements AsyncReadAt and AsyncGetSize for
FileProxy.

Bug: 68929

Change-Id: I92a3bdacc93efaf1b072e954abb80953589b12be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488023
Fuchsia-Auto-Submit: Sen Jiang <senj@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Sen Jiang <senj@google.com>