[rust] Remove clock ID from Timer.
zx_timer_create only accepts a clock ID of monotonic and when we remove
the zx_clock_get syscall the concept of defining a clock ID is going to
be a bit difficult to explain. Simplest just to remove it from the rust
API surface now.

Specifically this CL also means we can remove ClockID from Rust when we
remove the deprecated zx::Time::Get function.

Change-Id: I64b9fa9d67d4ab96fdea14a9877573b901c02714
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491542
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>