[fx] add fastboot tool to bringup
Some flash workflows depend on fastboot, which is copied from prebuilt
by the //build/images:fastboot GN target. This CL adds this target to
bringup so that the tool is copied without requiring special '--with'
arguments, and added to the host tool metadata file (tool_paths.json).

Note that it is not enough to add it to the host labels, like
//tools/devshell:fx, because the build_api_module walks the
metadata only on the $default_toolchain, and thus the tool would not be
expressed in tool_paths.json as it should be.

Fixed: 71094
Change-Id: I80948fc04a80c92dd2d328370b646126ceb5926e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492641
Reviewed-by: James Tucker <raggi@google.com>
Fuchsia-Auto-Submit: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>