[time] Cease verifying the kernel UTC clock.
This is one step towards removing the zx_clock_get and zx_clock_adjust
syscalls as per RFC-0008.

Running kcounters on a test device shows we've successfully removed all
users of the kernel UTC clock, except for the calls to monitor that clock.

The CL removes that monitoring along with the deprecated calls we
needed in fuchsia-zircon to maintain it. This allows us to finally cease
setting the kernel UTC in the next step.

Bug: 64063, 60521
Change-Id: I1ac0fe2524725719d6635347663d7d9ae713d0f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490584
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>