[component_manager_for_tests] Log to syslog in test
Tests explicitly set up the kernel logger. Although this is appropriate
for prod, we should not be doing this in the test because we want log
output to be reported with the test results, like other tests.

Also, remove a stale `with_unit_tests = true` on component_manager_bin
(the tests are declared in separate components).

Change-Id: If2883c898256d60524728e06ccb5cd7576d66eec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487267
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Gary Bressler <geb@google.com>