[fs] Move ulib/fs to //src/lib
Moves //zircon/system/ulib/fs to //src/lib/storage/vfs/cpp to conform to
the new directory layout and be next to the Rust VFS implementation.

There are several hundred files referencing the old location so
forwarding BUILD files and headers are left in the old location. The
users of this library will be updated in a second pass.

Bug: 70795
Change-Id: I614324874f4c6cc6e010f2b8d7b79afe3e3aba17
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490973
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Adam Barth <abarth@google.com>