[go] Add vendor/modules.txt
This is a follow up for http://fxrev.dev/490407. The native Go toolchain
picks up on /vendor/ and hard complains about lack of packages there,
even when modules.txt is not present. /vendor/ exists for Rust code at
the moment, and this makes it used for Go code too.

- Created a /vendor/modules.txt with: go mod vendor -e
- Synced with jiri update, as the go mod vendor command zaps out the
  /vendor/ directory.
- Updated ~/.gitignore to fix the vendor pattern to only ignore
  subdirectories under /vendor/. This permits removing 'negative'
  patterns that made the file more complicated than necessary.

This this change, IDE's "go to definition" functionality should work again.

A known workaround is to set the environment variable
GOFLAGS=-mod=readonly

This change should not affect the GN build.

Bug: 35565, 58804
Change-Id: I21e93302fd12da850069e172693457c163127d9e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493338
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Marc-Antoine Ruel <maruel@google.com>