[zedmon_test] Require -fake_zedmon_relpath flag
Otherwise, if the flag is unset the test will try to run an empty
command and hang until the test timeout is reached.

Long-term it might make sense to rewrite the fake_zedmon tool in Bash
instead of Go so that it can be run in the test without first doing a
build.

Change-Id: Ibab114571f90bf2979d16c95afe5760ba3707647
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493441
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>