[gidl] Decode vector / array tests for handle rights
This CL checks the decode path for some known corner cases when handle
rights and types are checked in vectors. Specifically, in some bindings
there is no place to store rights and type information on the objects
themselves so they need to be passed down in another way.

Change-Id: I6a18708474d3766f3ea4ed2b5062a7512c3f0f43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493278
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>