[host-target-testing] Support old and new-style nodenames
In http://fxrev.dev/481622, we changed how nodenames are computed. This
unfortunately doesn't work well with testing with builds that are older
than this CL. To fix this, this:

* creates a DeviceFinder type.
* extend device.Resolver to return multiple names
* change Device.Pave to wait for any of the device names to come online.
* add a `--old-device-name`, which defaults to $FUCHSIA_OLD_NODENAME.

We can consider removing thsi functionality once we no longer need to
test with builds older than 2021-02-22, but we might want to keep this
code around in case we ever want to change node names again.

Bug: 71056

Change-Id: I67470aa1862699941d0c253754cddf137d6c0d03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492979
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>