[kernel][x86] Refactor swapgs bug logic
We pull it into lib/arch to make it phys-friendly to enable its use in
code-patching in a follow-up change.

The previous swapgs logic determined what was affected in terms of
microarchitecture and had been written before the official CVE. Upon
rereading the official briefings, we now see that we actually expect
~every Intel CPU to be affected (with no indication that this will ever
be fixed) and so we can get away with a simpler "is Intel?" check.

Bug: 67615
Change-Id: I75bf18c2eaa92e77a4851acb2837a5932d23a766
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487867
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>