[bt][at] Round trip tests for AT command serde.
Add tests to confirm that serialization and deseserialization.  These
confirm that AT commands and responses can be converted to strings and
back--and vice versa--through several different paths, and the results
are the same.

Also fixes a bug where the parsers for AT commands were trying to read
from an empty iterator, causing intermittent crashes.

Change-Id: I052900e838cf85b518f815ad983485d95f9939bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487880
Commit-Queue: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>