[kernel][phys] Add base code for library handling x86_64 page tables.
This CL is the first of several to allow physboot to construct page
tables for both itself (bootstrapping) and the kernel proper (loading).

This initial CL only adds basic definitions for x86_64 page tables,
and some simple map/lookup functions. Upcoming CLs will improve support,
both in providing a public API for page table creation, support for
attributes (such as caching, permissions), and support for ARM page
tables.

In this CL, we represent the x86_64 hardware page table data structures
using the hwreg library. x86_64 contains 4 or 5 levels of page tables,
which have _mostly_ (but not completely) the same definitions. We've
opted to use a single structure to represent page table entries at all
levels, with custom getters/setters for the exceptions. Two possible
alternatives would be to try and represent each level as its own struct,
or to try and templatize the level. Neither approach was satisfactory,
as it forces code attempting to use the data structures less generic.

Bug: 67632
Change-Id: I8f214dc5bf346741244ce136587b6374b6b9d197
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486530
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>